[UPDATE] 파라미터 자료형 수정

+++ admin/src/main/java/kr/hs/entrydsm/admin/usecase/ExcelService.java
@@ -3,6 +3,6 @@ package kr.hs.entrydsm.admin.usecase;
 import java.io.IOException;
 
 public interface ExcelService {
-    void createAdmissionTicket(int receiptCode) throws IOException;
+    void createAdmissionTicket(long receiptCode) throws IOException;
     void createApplicantInformation();
 }
